added Rem function
