+++ sympy/__init__.py
@@ -53,7 +53,7 @@ from .core import (sympify, SympifyError, cacheit, Basic, Atom,
         Wild, Dummy, symbols, var, Number, Float, Rational, Integer,
         NumberSymbol, RealNumber, igcd, ilcm, seterr, E, I, nan, oo, pi, zoo,
         AlgebraicNumber, comp, mod_inverse, Pow, integer_nthroot, integer_log,
-        Mul, prod, Add, Mod, Rem, Rel, Eq, Ne, Lt, Le, Gt, Ge, Equality,
+        Mul, prod, Add, Mod, Rel, Eq, Ne, Lt, Le, Gt, Ge, Equality,
         GreaterThan, LessThan, Unequality, StrictGreaterThan, StrictLessThan,
         vectorize, Lambda, WildFunction, Derivative, diff, FunctionClass,
         Function, Subs, expand, PoleError, count_ops, expand_mul, expand_log,
@@ -110,7 +110,7 @@ from .series import (Order, O, limit, Limit, gruntz, series, approximants,
 from .functions import (factorial, factorial2, rf, ff, binomial,
         RisingFactorial, FallingFactorial, subfactorial, carmichael,
         fibonacci, lucas, motzkin, tribonacci, harmonic, bernoulli, bell, euler,
-        catalan, genocchi, partition, sqrt, root, Min, Max, Id, real_root,
+        catalan, genocchi, partition, sqrt, root, Min, Max, Id, real_root, Rem,
         cbrt, re, im, sign, Abs, conjugate, arg, polar_lift,
         periodic_argument, unbranched_argument, principal_branch, transpose,
         adjoint, polarify, unpolarify, sin, cos, tan, sec, csc, cot, sinc,
@@ -261,7 +261,7 @@ __all__ = [
     'Rational', 'Integer', 'NumberSymbol', 'RealNumber', 'igcd', 'ilcm',
     'seterr', 'E', 'I', 'nan', 'oo', 'pi', 'zoo', 'AlgebraicNumber', 'comp',
     'mod_inverse', 'Pow', 'integer_nthroot', 'integer_log', 'Mul', 'prod',
-    'Add', 'Mod','Rem', 'Rel', 'Eq', 'Ne', 'Lt', 'Le', 'Gt', 'Ge', 'Equality',
+    'Add', 'Mod', 'Rel', 'Eq', 'Ne', 'Lt', 'Le', 'Gt', 'Ge', 'Equality',
     'GreaterThan', 'LessThan', 'Unequality', 'StrictGreaterThan',
     'StrictLessThan', 'vectorize', 'Lambda', 'WildFunction', 'Derivative',
     'diff', 'FunctionClass', 'Function', 'Subs', 'expand', 'PoleError',
@@ -327,7 +327,7 @@ __all__ = [
     'factorial', 'factorial2', 'rf', 'ff', 'binomial', 'RisingFactorial',
     'FallingFactorial', 'subfactorial', 'carmichael', 'fibonacci', 'lucas',
     'motzkin', 'tribonacci', 'harmonic', 'bernoulli', 'bell', 'euler', 'catalan',
-    'genocchi', 'partition', 'sqrt', 'root', 'Min', 'Max', 'Id', 'real_root',
+    'genocchi', 'partition', 'sqrt', 'root', 'Min', 'Max', 'Id', 'real_root', 'Rem',
     'cbrt', 're', 'im', 'sign', 'Abs', 'conjugate', 'arg', 'polar_lift',
     'periodic_argument', 'unbranched_argument', 'principal_branch',
     'transpose', 'adjoint', 'polarify', 'unpolarify', 'sin', 'cos', 'tan',
