+++ sympy/functions/elementary/miscellaneous.py
@@ -1,4 +1,4 @@
-from sympy.core import Function, S, sympify
+from sympy.core import Function, S, sympify, NumberKind
 from sympy.utilities.iterables import sift
 from sympy.core.add import Add
 from sympy.core.containers import Tuple
@@ -865,3 +865,55 @@ class Min(MinMaxBase, Application):
 
     def _eval_is_negative(self):
         return fuzzy_or(a.is_negative for a in self.args)
+
+
+class Rem(Function):
+    """
+    Parameters
+    ==========
+
+    p : Expr
+        Dividend.
+
+    q : Expr
+        Divisor.
+
+    Notes
+    =====
+
+    the remainder always has the same sign as the dividend.
+
+    Examples
+    ========
+
+    >>> from sympy import Rem
+    >>> from sympy.abc import x, y
+    >>> Rem(x**3, y)
+    Rem(x**3, y)
+    >>> _.subs({x: -5, y: 3})
+    -2
+
+    """
+
+    kind = NumberKind
+
+    @classmethod
+    def eval(cls,p,q):
+        def doit(p,q):
+            """ the function remainder if both p,q are numbers
+                and q is not zero
+            """
+
+            if q.is_zero:
+                raise ZeroDivisionError("Division by zero")
+            if p is S.NaN or q is S.NaN or p.is_finite is False or q.is_finite is False:
+                return S.NaN
+            if p is S.Zero or p in (q, -q) or (p.is_integer and q == 1):
+                return S.Zero
+
+            if q.is_Number:
+                if p.is_Number:
+                    return p - int(p/q)*q
+        rv = doit(p, q)
+        if rv is not None:
+            return rv
